refactor: dnd 이벤트핸들러 모듈화
